backend/requirements.txt
Django==4.2.*
djangorestframework==3.15.*   # DRF included for later steps—harmless now
elasticsearch==8.12.*
gunicorn==20.1.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
pdfminer.six==20250506
tqdm==4.67.1